% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssmc_data_description.R
\docType{data}
\name{potential.dat}
\alias{potential.dat}
\title{Example potential site data}
\format{A 25 x 3 data frame}
\usage{
data(potential.dat)
}
\description{
Sample file to be used with \code{\link[ssmc]{best.locale}}
}
\details{
A data frame containg the necessary components to use the \code{\link[ssmc]{best.locale}} function

\itemize{
   \item site - Numeric site ID
   \item x -  x-coordinate
   \item y -  y-coordinate
   }
}
\keyword{datasets}

