% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/site.analysis.r
\name{site.analysis}
\alias{site.analysis}
\title{A function to assess average source-sink status of populations as well as rank importance of populations in contributing to the metapopulation}
\usage{
site.analysis(sites,
pop.abun,
met.size=NULL,
prop.philo=.85,
sd.philo = 0.05,
lower.upper_philo=c(lower=0, upper=1),
prop.survive=0.2,
sd.survive = 0.075,
lower.upper_survive = c(lower=0, upper=1),
dispersal,
sd.dispersal,
lower.upper_dispersal = c(lower=1, upper=Inf),
iterations,
seed=NULL)
}
\arguments{
\item{sites}{A three-column matrix or data frame providing the name of the site (column 1) and the xy coordinates of current sites (columns 2 & 3)}

\item{pop.abun}{A two-column data frame with mean population size in column 1 and standard deviation in column 2}

\item{met.size}{(Default = NULL) If specified, must be a two-column data frame with the mean (column 1) and standard deviation (column 2) of the size of metamoprhs or late-stage larvae. See Details for more information concerning this parameter.}

\item{prop.philo}{(Default = 0.85) Mean proportion of population that are philopatric to their natal population}

\item{sd.philo}{(Default = 0.05) Standard deviation of proportion of population that are philopatric to their natal population}

\item{lower.upper_philo}{(Default = c(lower=0, upper=1)) Threshold lower and upper values for the proportion of philopatric individuals. Must be provided as a two-element vector with lower value first. See Details for use}

\item{prop.survive}{(Default = 0.2) Mean survival of to adulthood. Ignored if \code{met.size} is specified. See Details for more information.}

\item{sd.survive}{(Default = 0.075) Standard deviation of survival of to adulthood.}

\item{lower.upper_survive}{(Default = c(lower=0, upper=1))Threshold lower and upper values for the proportion of individuals surviving to adulthood. Must be provided as a two-element vector with lower value first. See Details}

\item{dispersal}{Mean dispersal distance. Should be specified in units meaningful to the coordinate system describing the spatial location of populations. See Details for how dispersal is estimated.}

\item{sd.dispersal}{Standard deviation of dispersal distance.}

\item{lower.upper_dispersal}{(Default = c(lower=1, upper=Inf))Threshold lower and upper values for average dispersal distance. Must be provided as a two-element vector with minimum value first. See Details}

\item{iterations}{Number of Monte Carlo iterations to run}

\item{seed}{Optional to set the seed for repeatability  among model runs (Default = NULL)}
}
\value{
This function returns a list with two components: (1) \code{$summary.df} is a data frame with the results averaged over all Monte Carlo iterations, and (2) \code{$results.list} is a list of length equal to the number of Monte Carlo iterations and contains the raw results from the analysis.

Elements of the summary data frame include: \cr
(1) \code{immig}: The average number of immigrants entering each population; \cr
(2) \code{philo}: The average number of philopatric individuals; \cr
(3) \code{emig}: The average number of emigrants dispersing from a population; \cr
(4) \code{pct_immig}: Average percentage of a population that is comprised of immigrants; \cr
(5) \code{pct_sink}: The frequency that a population acted as a sink (i.e. immigrants > philopatric) across all simulations; \cr
(6) \code{pct_src}: The frequency that a population acted as a source (i.e. immigrant <= philopatric) across all simulations; \cr
(7) \code{cont_lamb}: The contribution of each location to the metapopulation; \cr
(8) \code{rank_cont}: The rank order importance of each location to the metapopulation; \cr
(9) \code{delta_lam}: The change in the value of the the leading eigen value of the landscape matrix when each population is removed, which is interpreted as the metapopulation capacity. This is calculated as the log ratio of values [delta_lam = log(lambda.new / lambda.old)]; \cr
(10) \code{rank_lamb}: The rank order importance of each location to metapopulation capacity; \cr
(11) \code{avg_rank}: The rank importance of each potential location, determined by averaging \code{rank_cont} and \code{rank_lamb}.
}
\description{
A function to assess average source-sink status of populations as well as rank importance of populations in contributing to the metapopulation
}
\details{
If \code{met.size} is specified, the probability of surviving to adulthood is determined using the equation:\cr
logit(p.survive) =  -1.366 + 0.87 * size\cr
This equation comes from Altwegg & Reyer (2003). Mean and standard deviation values for \code{met.size} must be reported in standard units such that the mean and standard deviation of observations equal zero and one, respectively (i.e. scale and center observations). In the absence of metamorph, survival probability and variation can be specified using \code{prop.survive}.

This model assumes uncertainty or variability in (1) population size; \cr \cr (2) size of metamorphs OR proportion surviving; \cr \cr (3) proportion of population that is philopatric; \cr \cr (4) mean dispersal distance. Uncertainty in these parameters is incorporated through repeated draws from normal distibutions with a mean and standard deviation as specified. Because some values are unrealistic (e.g., survival > 1), a truncated normal distribution is used, which requires the specification of lower and upper values. If there are no limits on the lower or upper values, then \code{-Inf} or \code{Inf} should be specified. Lower and upper values must be provided as a two-element vector (e.g., c(0,1)) for \code{lower.upper_philo}, \code{lower.upper_survive}, and \code{lower.upper_dispersal}

Probability of dispersal between two populations is determined using an incidence function wherein the probability of connectivity is a negative exponential relationship with 1/mean dispersal controlling the rate of decay.
}
\examples{
# Assess existing populations
   site_results <- site.analysis(sites = site.dat[,1:3],
   pop.abun = site.dat[,4:5],
   met.size = site.dat[,6:7],
   prop.philo = 0.95,
   sd.philo = 0.05,
   lower.upper_philo=c(lower=0, upper=1),
   # prop.survive=0.2,      ## Not needed b/c met.size is specified
   # sd.survive = 0.075,    ## Not needed b/c met.size is specified
   lower.upper_survive = c(lower=0, upper=1),
   dispersal = 25,
   sd.dispersal = 10,
   lower.upper_dispersal = c(lower=10, upper=Inf),
   iterations = 10,
   seed = 123)
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
\references{
Altwegg, R., and H.-U. Reyer. 2003. Patterns of natural selection on size at metamorphosis in water frogs. Evolution 57:872-882.
}
\seealso{
\code{\link[ssmc]{ssmc_summary}}
}

