% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmc_data_description.R
\docType{package}
\name{ssmc-package}
\alias{ssmc-package}
\title{About this R package}
\description{
This package contains two functions: (1) Assess the average source-sink status of existing populations as well as their rank importance to the metapopulation; (2) Determine the optimal location for habitat creation that will provide the greatest benefit to the metapopulation.
}
\details{
\tabular{ll}{
   Package: \tab ssmc\cr
   Type: \tab Package\cr
   License: \tab >=GPL-2\cr
 }
 This package provides functions to conduct Monte Carlo simulations of demographic connectivity models. These functions can assess the contributions and importance of existing populations and can also assess the potential contributions of populations following habitat creation or restoration.
}
\author{
Bill Peterman \email{Peterman.73@osu.edu}
}
\references{
Please cite:
Manuscript is being prepared for publication
}

