% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sort_results.R
\name{ssmc_summary}
\alias{ssmc_summary}
\title{A function to extract the summary data from \code{\link[ssmc]{site.analysis}} or \code{\link[ssmc]{best.locale}} functions}
\usage{
ssmc_summary(ssmc_results, ...)
}
\arguments{
\item{ssmc_results}{Output object from \code{\link[ssmc]{site.analysis}} or \code{\link[ssmc]{best.locale}}}

\item{...}{Optional to select which column to sort by when summarizing the results data frame. The default is to sort by \code{rank}}
}
\value{
This function returns a data frame of the summarized results from either the \code{\link[ssmc]{site.analysis}} or \code{\link[ssmc]{best.locale}} sorted by the specified column.
}
\description{
A function to extract the summary data from \code{\link[ssmc]{site.analysis}} or \code{\link[ssmc]{best.locale}} functions
}
\examples{
   # Assess potential new locations
   best_results <- best.locale(sites = site.dat[,1:3],
   potential.sites = potential.dat,
   pop.abun = site.dat[,4:5],
   met.size = site.dat[,6:7],
   prop.philo = 0.95,
   sd.philo = 0.05,
   min.max_philo=c(min=0, max=1),
   # prop.survive=0.2,      ## Not needed b/c met.size is specified
   # sd.survive = 0.075,    ## Not needed b/c met.size is specified
   min.max_survive = c(min=0, max=1),
   dispersal = 25,
   sd.dispersal = 10,
   min.max_dispersal = c(min=10, max=Inf),
   iterations = 10,
   seed = 123)

## Default
   best_locales <- ssmc_summary(best_results)

## Sort by the adjusted rank
   best_locales <- ssmc_summary(best_results, 'rank_adj')
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
